create new structure and create a system to manage different protocols
