+++ block_info.cc
@@ -1,0 +1,20 @@
+/*
+  Ghost File System, or simply GhostFS
+  Copyright (C) 2016 Raphael S. Carvalho
+
+  This program can be distributed under the terms of the GNU GPL.
+  See the file COPYING.
+*/
+
+#include "block_info.h"
+
+
+void block_info::reset() {
+    _present = false;
+    _blk = nullptr;
+}
+
+void block_info::set_block(block *blk) {
+    _present = true;
+    _blk = blk;
+}
