+++ utils.h
@@ -1,0 +1,17 @@
+/*
+  Ghost File System, or simply GhostFS
+  Copyright (C) 2016 Raphael S. Carvalho
+
+  This program can be distributed under the terms of the GNU GPL.
+  See the file COPYING.
+*/
+
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <string>
+#include <vector>
+
+std::vector<std::string> split(const std::string &s, char delim);
+
+#endif // UTILS_H
